requirements.txt
Telethon>=1.34.0
FastAPI>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
websockets>=12.0
